Reduce the number of files
