+++ .gitignore
@@ -2,3 +2,6 @@ bin
 obj
 .objs
 .mgit
+*.d
+*.o
+
