--- cmd_cat_file.h
@@ -1,8 +1,0 @@
-#ifndef MGIT_CMD_CAT_FILE
-#define MGIT_CMD_CAT_FILE
-
-#include "types.h"
-
-cmd_handler cmd_cat_file;
-
-#endif // MGIT_CMD_CAT_FILE
