--- cmd_hash_object.h
@@ -1,8 +1,0 @@
-#ifndef MGIT_CMD_HASH_OBJECT_H
-#define MGIT_CMD_HASH_OBJECT
-
-#include "types.h"
-
-cmd_handler cmd_hash_object;
-
-#endif // MGIT_CMD_HASH_OBJECT_H
