--- cmd_init.h
@@ -1,8 +1,0 @@
-#ifndef MGIT_INIT_H
-#define MGIT_INIT_H
-
-#include "types.h"
-
-cmd_handler cmd_init;
-
-#endif // MGIT_INIT_H
