--- cmd_write_tree.c
@@ -1,4 +1,0 @@
-
-#include "cmd_write_tree.h"
-
-int cmd_write_tree(int argc, const char *argv[]) { return MGIT_SUCCESS; }
