--- cmd_write_tree.h
@@ -1,8 +1,0 @@
-#ifndef MGIT_CMD_WRITE_TREE_H
-#define MGIT_CMD_WRITE_TREE_H
-
-#include "types.h"
-
-cmd_handler cmd_write_tree;
-
-#endif // MGIT_CMD_WRITE_TREE_H
