--- mgit.cbp
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
-<CodeBlocks_project_file>
-	<FileVersion major="1" minor="6" />
-	<Project>
-		<Option title="mgit" />
-		<Option pch_mode="2" />
-		<Option compiler="gcc" />
-		<Build>
-			<Target title="Debug">
-				<Option output="bin/Debug/mgit" prefix_auto="1" extension_auto="1" />
-				<Option object_output="obj/Debug/" />
-				<Option type="1" />
-				<Option compiler="gcc" />
-				<Option parameters="hash-object mgit.c" />
-				<Compiler>
-					<Add option="-g" />
-				</Compiler>
-				<Linker>
-					<Add library="/lib/libcrypto.so" />
-					<Add library="/lib/libssl.so" />
-				</Linker>
-			</Target>
-		</Build>
-		<Compiler>
-			<Add option="-Wall" />
-		</Compiler>
-		<Unit filename="argparse.c">
-			<Option compilerVar="CC" />
-		</Unit>
-		<Unit filename="cmd_cat_file.c">
-			<Option compilerVar="CC" />
-		</Unit>
-		<Unit filename="cmd_cat_file.h" />
-		<Unit filename="cmd_hash_object.c">
-			<Option compilerVar="CC" />
-		</Unit>
-		<Unit filename="cmd_hash_object.h" />
-		<Unit filename="cmd_init.c">
-			<Option compilerVar="CC" />
-		</Unit>
-		<Unit filename="cmd_init.h" />
-		<Unit filename="filesystem.c">
-			<Option compilerVar="CC" />
-		</Unit>
-		<Unit filename="filesystem.h" />
-		<Unit filename="mgit.c">
-			<Option compilerVar="CC" />
-		</Unit>
-		<Unit filename="objects.c">
-			<Option compilerVar="CC" />
-		</Unit>
-		<Unit filename="objects.h" />
-		<Unit filename="subcommands.h" />
-		<Unit filename="types.h" />
-		<Extensions>
-			<AutoVersioning>
-				<Scheme minor_max="10" build_max="0" rev_max="0" rev_rand_max="10" build_times_to_increment_minor="100" />
-				<Settings autoincrement="1" date_declarations="1" use_define="0" update_manifest="1" do_auto_increment="0" ask_to_increment="0" language="C++" svn="0" svn_directory="" header_path="version.h" />
-				<Changes_Log show_changes_editor="0" app_title="released version %M.%m.%b of %p" changeslog_path="ChangesLog.txt" />
-				<Code header_guard="VERSION_H" namespace="AutoVersion" prefix="" />
-			</AutoVersioning>
-		</Extensions>
-	</Project>
-</CodeBlocks_project_file>
