--- mgit.cscope_file_list
@@ -1,13 +1,0 @@
-"/home/felix/proj/mgit/subcommands.h"
-"/home/felix/proj/mgit/filesystem.c"
-"/home/felix/proj/mgit/cmd_hash_object.c"
-"/home/felix/proj/mgit/filesystem.h"
-"/home/felix/proj/mgit/cmd_init.h"
-"/home/felix/proj/mgit/cmd_init.c"
-"/home/felix/proj/mgit/types.h"
-"/home/felix/proj/mgit/objects.h"
-"/home/felix/proj/mgit/argparse.c"
-"/home/felix/proj/mgit/version.h"
-"/home/felix/proj/mgit/mgit.c"
-"/home/felix/proj/mgit/objects.c"
-"/home/felix/proj/mgit/cmd_hash_object.h"
