--- mgit.depend
@@ -1,62 +1,0 @@
-# depslib dependency file v1.0
-1602388427 source:/home/felix/proj/mgit/argparse.c
-
-1602446420 source:/home/felix/proj/mgit/cmd_init.c
-	<stdio.h>
-
-1602343437 source:/home/felix/proj/mgit/main.c
-	<stdio.h>
-	<stdlib.h>
-
-1602647077 source:/home/felix/proj/mgit/mgit.c
-	"subcommands.h"
-	<string.h>
-	<stdio.h>
-
-1602644934 /home/felix/proj/mgit/subcommands.h
-	"cmd_init.h"
-	"cmd_hash_object.h"
-	"cmd_cat_file.h"
-
-1602387837 /home/felix/proj/mgit/init.h
-	"types.h"
-
-1602646106 /home/felix/proj/mgit/types.h
-	<stdio.h>
-
-1602646816 source:/home/felix/proj/mgit/objects.c
-	"objects.h"
-	"filesystem.h"
-	<stdio.h>
-	<stdlib.h>
-	<string.h>
-	<openssl/sha.h>
-
-1602646611 /home/felix/proj/mgit/objects.h
-	"types.h"
-
-1602435488 source:/home/felix/proj/mgit/filesystem.c
-	<sys/types.h>
-	<sys/stat.h>
-
-1602646828 /home/felix/proj/mgit/filesystem.h
-
-1602644934 /home/felix/proj/mgit/cmd_init.h
-	"types.h"
-
-1602644934 /home/felix/proj/mgit/cmd_hash_object.h
-	"types.h"
-
-1602646776 source:/home/felix/proj/mgit/cmd_hash_object.c
-	"cmd_hash_object.h"
-	"objects.h"
-	"types.h"
-	<stdio.h>
-
-1602646789 source:/home/felix/proj/mgit/cmd_cat_file.c
-	"cmd_cat_file.h"
-	"objects.h"
-
-1602644934 /home/felix/proj/mgit/cmd_cat_file.h
-	"types.h"
-
